refactor(cliente): mover los tipos al directorio types
Se pueden necesitar los mismos tipos en otras partes de la aplicación
