+++ client/src/types/comun-tipos.ts
