+++ client/src/types/mensaje-tipos.ts
